+++ packages/plugin-pnpm/sources/PnpmLinker.ts
@@ -1,8 +1,7 @@
-import {Descriptor, FetchResult, formatUtils, Installer, InstallPackageExtraApi, Linker, LinkOptions, LinkType, Locator, LocatorHash, Manifest, MessageName, MinimalLinkOptions, Package, Project, structUtils} from '@yarnpkg/core';
-import {Dirent, Filename, PortablePath, ppath, xfs}                                                                                                                                                             from '@yarnpkg/fslib';
-import {jsInstallUtils}                                                                                                                                                                                         from '@yarnpkg/plugin-pnp';
-import {UsageError}                                                                                                                                                                                             from 'clipanion';
-import pLimit                                                                                                                                                                                                   from 'p-limit';
+import {Descriptor, FetchResult, formatUtils, Installer, InstallPackageExtraApi, Linker, LinkOptions, LinkType, Locator, LocatorHash, Manifest, MessageName, MinimalLinkOptions, Package, Project, miscUtils, structUtils} from '@yarnpkg/core';
+import {Dirent, Filename, PortablePath, ppath, xfs}                                                                                                                                                                        from '@yarnpkg/fslib';
+import {jsInstallUtils}                                                                                                                                                                                                    from '@yarnpkg/plugin-pnp';
+import {UsageError}                                                                                                                                                                                                        from 'clipanion';
 
 export type PnpmCustomData = {
   pathByLocator: Map<LocatorHash, PortablePath>;
@@ -72,7 +71,7 @@ export class PnpmLinker implements Linker {
 }
 
 class PnpmInstaller implements Installer {
-  private asyncActions = new AsyncActions();
+  private readonly asyncActions = new miscUtils.AsyncActions(10);
 
   constructor(private opts: LinkOptions) {
     // Nothing to do
@@ -389,59 +388,5 @@ async function removeIfEmpty(dir: PortablePath) {
     if (error.code !== `ENOENT` && error.code !== `ENOTEMPTY`) {
       throw error;
     }
-  }
-}
-
-type Deferred = {
-  promise: Promise<void>;
-  resolve: () => void;
-  reject: (err: Error) => void;
-};
-
-function makeDeferred(): Deferred {
-  let resolve: () => void;
-  let reject: (err: Error) => void;
-
-  const promise = new Promise<void>((resolveFn, rejectFn) => {
-    resolve = resolveFn;
-    reject = rejectFn;
-  });
-
-  return {promise, resolve: resolve!, reject: reject!};
-}
-
-class AsyncActions {
-  deferred = new Map<string, Deferred>();
-  promises = new Map<string, Promise<void>>();
-  limit = pLimit(10);
-
-  set(key: string, factory: () => Promise<void>) {
-    let deferred = this.deferred.get(key);
-    if (typeof deferred === `undefined`)
-      this.deferred.set(key, deferred = makeDeferred());
-
-    const promise = this.limit(() => factory());
-    this.promises.set(key, promise);
-
-    promise.then(() => {
-      if (this.promises.get(key) === promise) {
-        deferred!.resolve();
-      }
-    }, err => {
-      if (this.promises.get(key) === promise) {
-        deferred!.reject(err);
-      }
-    });
-
-    return deferred.promise;
-  }
-
-  reduce(key: string, factory: (action: Promise<void>) => Promise<void>) {
-    const promise = this.promises.get(key) ?? Promise.resolve();
-    this.set(key, () => factory(promise));
-  }
-
-  async wait() {
-    await Promise.all(this.promises.values());
   }
 }
